Begun refactoring data processing from animation and signal analyzer classes
